move `ActionType` to action_type.go
